wizard: Extract constants and normalize curl arguments
The use of the curl_args variable was inconsistent: Some flags would be
specified in the array whereas others directly in the curl command as
either short or long flags. They are now all long flags in the array.

Software versions were extracted into script variables. No effort was
made to guarantee we can simply update these variables to increase the
value; In fact, it's probably dangerous to do so, without considering
an upgrade path. This however, land the groundwork for such feature
later.

Fix impurity on configure_mongodb_repo, which would use /etc/os-release
instead of the facts previously collected.

Signed-off-by: Roosembert Palacios <roos@be-bop.io>