wizard: Ensure mongodb rs is configured when externally installed
Previously, an external installation of mongodb would require the user
to also configure the replica-set. Most users will likely just ask an
LLM “How do I install mongodb in XXX”; thus the risk of the cluster
being configured without a RS is non-null.

The cluster is now configured in step initialize_mongodb_rs. Thus even
if MongoDB was installed with the incorrect configurationm, the task
will correctly initialize the cluster.

Signed-off-by: Roosembert Palacios <roos@be-bop.io>