wizard: Enable detection of outdated configuration files
The configuration file now contains a “fingerprint” of the “template”
used by the wizard to generate the configuration file. This is an opaque
identifier and enables the wizard to determine whether the configuration
file should be rewritten.
Mind that the template fingerprint lives a separate life than the wizard
version, as the wizard may be updated without changing the template.

When the config file is replaced, all user customization is stiched into
the new file.

Signed-off-by: Roosembert Palacios <roos@be-bop.io>