Bundle wizard into a packaged release
This commit adds be-bop-bootstrap, a self-contained, self-extracting
bundle. the bootstrap bundles the wizard and the cli; when run, the bootstrap
extracts all bundled artifacts and runs the wizard.

This enables shipping matching versions between the wizard and cli.

In the future, the same mechanism can be used to vendor pre-seeded (or
offline) installs.

Signed-off-by: Roosembert Palacios <roos@be-bop.io>