Bundle wizard into a packaged release
This commit adds be-bop-bootstrap, a self-contained, self-extracting
bundle. the bootstrap bundles the wizard and the cli; when run, the bootstrap
extracts all bundled artifacts and runs the wizard.

There's currently one breaking change with the wizard invocation: An
extra `--` argument needs to be specified **before** the wizard flags:
e.g. `be-bop-bootstrap -- --help`.

This enables shipping matching versions between the wizard and cli.

In the future, the same mechanism can be used to vendor additional
assets or artifacts for pre-seeded (or offline) installations.

Signed-off-by: Roosembert Palacios <roos@be-bop.io>