Have wizard use be-bop-cli to install the latest release
The pnpm installation step was refactored into enabling corepack
instead. pnpm is still enabled opportunistically, but be-bop-cli
ensures be-BOP runs with the pinned version.

A sudoers rule is installed alongside the cli to enable be-bop-cli to
manage the bebop systemd service. Mind that polkit cannot be used here
since it's now always available (e.g. in containers).

Signed-off-by: Roosembert Palacios <roos@be-bop.io>